2006-06-25  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com> 	* src/Util.cc 	(difftv): Added a cast to (long long int). 	* src/TorrentMan.cc 	(advertisePiece): Use push_front. 	(getAdvertisedPieceIndexes): A performance improvement was made. 	* src/Time.h 	(getTimeInMicros): Added a cast to (long long int). 	(getTimeInMillis): Added a cast to (long long int).

+++ src/Util.cc
@@ -94,7 +94,7 @@ long long int Util::difftv(struct timeval tv1, struct timeval tv2) {
   if(tv1.tv_sec < tv2.tv_sec || tv1.tv_sec == tv2.tv_sec && tv1.tv_usec < tv2.tv_usec) {
     return 0;
   }
-  return ((tv1.tv_sec-tv2.tv_sec)*1000000+
+  return ((long long int)(tv1.tv_sec-tv2.tv_sec)*1000000+
 	  tv1.tv_usec-tv2.tv_usec);
 }
